// generated/src/app/api-models/aucls-model-determinant.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuclsModelDeterminant {
  auclsModelId: string;
  authsColumnName: string;
  operator: string;
  fromValue: string;
  thruValue: string;
  authsColumnTypeCode: number;
  columnOccurrence: number;
  columnType: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}